fix(types): Enable plugin to be imported under ES6 (#316)
Because the webpack plugin is written in vanilla JS (rather than TS), we have a manually-generated type declaration file, where we default-export the `SentryCliPlugin` class. We also have a file (`cjs.js`) which is meant to make the plugin play nicely with CommonJS imports. And if the user is indeed using CommonJS imports (`const SentryWebpackPlugin = require('@sentry/webpack-plugin')`), that all works great.

But if you try to do ES6 imports, the only form which gets you anything besides `undefined` is `import * as SentryWebpackPlugin from '@sentry/webpack-plugin`.* As far as TS is concerned, though, that creates a namespace rather than a class, and so it will refuse to let you call `new SentryWebpackPlugin`, instead telling you that `SentryWebpackPlugin` is not contructable. One solution is to use `esModuleInterop`, but that's [gotten us into trouble before](https://github.com/getsentry/sentry-javascript/issues/3401#issuecomment-826835211) and in fact breaks some of our own SDK code.

 *The other two options are `import SentryCliPlugin as SentryWebpackPlugin from '@sentry/webpack-plugin'` and `import { default as SentryWebpackPlugin } from '@sentry/webpack-plugin'`, both of which let you call `new` on them, but neither of which currently actually gives you anything other than `undefined`). 

The alternative (which is what's done in this change) is to do two things:

1) Use the older, more universal `exports = ...` syntax in our type declaration file rather than our current `export default ...` syntax (which, according to the TS docs, [actually won’t work](https://www.typescriptlang.org/docs/handbook/declaration-files/templates/module-d-ts.html#default-exports) unless `esModuleInterop` is set to `true`). This then necessitates moving the options interface into a namespace for the module.

2) Explicitly create an export under the name `default`, so that the two alternatives above actually come up with something rather than being undefined.

This shouldn't affect any current imports, as the first change just moves to an equivalent syntax, and the second change is purely additive.